requirements.txt
Pillow
uvicorn
fastapi
opencv-python
numpy
python-multipart
jinja2
transformers
torch